requirements.txt
torch~=2.1.2
torchvision~=0.16.2
thop~=0.1.1-2209072238
einops~=0.7.0
numpy~=1.26.4
opencv-python~=4.9.0.80
pillow~=10.0.1
tqdm~=4.65.0
scikit-image~=0.22.0